Refine and refactor the code
