+++ .gitattributes
@@ -1,0 +1,21 @@
+* text=auto
+*.css linguist-vendored
+*.scss linguist-vendored
+*.js linguist-vendored
+.dockerignore export-ignore
+.editorconfig export-ignore
+.phpcs.xml.dist export-ignore
+.styleci.yml export-ignore
+.travis.yml export-ignore
+phpinsights.php export-ignore
+phpstan.neon.dist export-ignore
+phpunit.xml.dist export-ignore
+CHANGELOG.md export-ignore
+CONTRIBUTING.md export-ignore
+README.md export-ignore
+RELEASE.md export-ignore
+.github export-ignore
+docker export-ignore
+docs export-ignore
+tests export-ignore
+art export-ignore
