+++ src/Readers/AbstractCsvReader.php
@@ -2,7 +2,6 @@
 
 namespace Phpcsv\CsvHelper\Readers;
 
-use Phpcsv\CsvHelper\Configs\CsvConfig;
 use Phpcsv\CsvHelper\Contracts\CsvConfigInterface;
 use Phpcsv\CsvHelper\Contracts\CsvReaderInterface;
 use SplFileObject;
@@ -19,73 +18,30 @@ abstract class AbstractCsvReader implements CsvReaderInterface
 
     protected ?int $recordCount = null;
 
-    public function __construct(
+    abstract public function __construct(
         ?string $source = null,
         ?CsvConfigInterface $config = null
-    ) {
-        $this->config = $config ?? new CsvConfig;
+    );
 
-        if ($source !== null) {
-            $this->setSource($source);
-        }
-    }
+    abstract public function getReader(): ?SplFileObject;
 
-    public function getReader(): ?SplFileObject
-    {
-        return $this->reader;
-    }
+    abstract public function getConfig(): CsvConfigInterface;
 
-    public function getConfig(): CsvConfigInterface
-    {
-        if (! isset($this->config)) {
-            $this->config = new CsvConfig;
-        }
+    abstract public function getRecordCount(): ?int;
 
-        return $this->config;
-    }
+    abstract public function rewind(): void;
 
-    public function getRecordCount(): ?int
-    {
-        return null;
-    }
+    abstract public function getCurrentPosition(): int;
 
-    public function rewind(): void
-    {
-        $this->position = 0;
-    }
+    abstract public function getRecord(): array|false;
 
-    public function getCurrentPosition(): int
-    {
-        return $this->position;
-    }
+    abstract public function getHeader(): array|false;
 
-    public function getRecord(): array|false
-    {
-        return false;
-    }
+    abstract public function hasRecords(): bool;
 
-    public function getHeader(): array|false
-    {
-        return false;
-    }
+    abstract public function setSource(string $source): void;
 
-    public function hasRecords(): bool
-    {
-        return false;
-    }
+    abstract public function getSource(): string;
 
-    public function setSource(string $source): void
-    {
-        $this->getConfig()->setPath($source);
-    }
-
-    public function getSource(): string
-    {
-        return $this->config->getPath();
-    }
-
-    public function setConfig(CsvConfigInterface $config): void
-    {
-        $this->config = $config;
-    }
+    abstract public function setConfig(CsvConfigInterface $config): void;
 }
