+++ src/Writers/AbstractCsvWriter.php
@@ -2,7 +2,6 @@
 
 namespace Phpcsv\CsvHelper\Writers;
 
-use Phpcsv\CsvHelper\Configs\CsvConfig;
 use Phpcsv\CsvHelper\Contracts\CsvConfigInterface;
 use Phpcsv\CsvHelper\Contracts\CsvWriterInterface;
 use SplFileObject;
@@ -15,36 +14,18 @@ abstract class AbstractCsvWriter implements CsvWriterInterface
 
     protected ?SplFileObject $writer = null;
 
-    public function __construct(
+    abstract public function __construct(
         ?string $target = null,
         ?CsvConfigInterface $config = null
-    ) {
-        $this->config = $config ?? new CsvConfig;
+    );
 
-        if ($target !== null) {
-            $this->setTarget($target);
-        }
-    }
+    abstract public function getWriter(): ?SplFileObject;
 
-    public function getWriter(): ?SplFileObject
-    {
-        return null;
-    }
+    abstract public function getConfig(): CsvConfigInterface;
 
-    public function getConfig(): CsvConfigInterface
-    {
-        return $this->config;
-    }
+    abstract public function write(array $data): void;
 
-    public function write(array $data): void {}
+    abstract public function setTarget(string $target): void;
 
-    public function setTarget(string $target): void
-    {
-        $this->config->setPath($target);
-    }
-
-    public function getTarget(): string
-    {
-        return $this->config->getPath();
-    }
+    abstract public function getTarget(): string;
 }
